bug fix, CollisionManager.cs pair removal
git-tfs-id: [http://tfs.bumc.bu.edu:8080/tfs/BU-MICRO-MSI]$/Daphne/Daphne-axin3;C6096

+++ Daphne/CollisionManager.cs
@@ -366,7 +366,7 @@ namespace Daphne
             // remove the pairs that got scheduled for removal
             if (removalPairKeys != null)
             {
-                foreach (int key in removalPairKeys)
+                foreach (long key in removalPairKeys)
                 {
                     pairs.Remove(key);
                 }
